+++ backend/src/main/resources/db/migration/V6__cria_tabela_agendamentos.sql
@@ -3,8 +3,8 @@ CREATE TABLE IF NOT EXISTS agendamentos (
     data_hora TIMESTAMPTZ NOT NULL,
     valor_total NUMERIC(10, 2) NOT NULL,
     tempo_duracao_em_minutos INTEGER NOT NULL,
-    barbearia_id INTEGER REFERENCES barbearias(id),
+    barbearia_id BIGINT REFERENCES barbearias(id),
     CONSTRAINT fk_barbearia FOREIGN KEY (barbearia_id) REFERENCES barbearias(id),
-    cliente_id INTEGER REFERENCES clientes(id),
+    cliente_id BIGINT REFERENCES clientes(id),
     CONSTRAINT fk_cliente FOREIGN KEY (cliente_id) REFERENCES clientes(id)
 );
